Split queries into cost and usage
